+++ app/src/main/java/com/example/felipe/criminalintent/Crime.kt
@@ -3,6 +3,9 @@ package com.example.felipe.criminalintent
 import java.util.UUID
 import java.util.Date
 
-data class Crime(var date: Date = Date(), var title: String? = null, var isSolved: Boolean = false, var isPoliceRequire: Boolean = false) {
+data class Crime(var date: Date = Date(),
+                 var title: String? = null,
+                 var isSolved: Boolean = false,
+                 var isPoliceRequire: Boolean = false) {
     val id: UUID = UUID.randomUUID()
 }
