+++ app/src/main/java/com/example/felipe/criminalintent/CrimeListActivity.kt
@@ -1,7 +1,5 @@
 package com.example.felipe.criminalintent
 
 class CrimeListActivity : SingleFragmentActivity() {
-    override fun initFragments() {
-        CrimeListFragment().addFragmentToFragmentContainer()
-    }
+    override fun initFragments() { CrimeListFragment().addFragmentToFragmentContainer() }
 }
