+++ app/src/main/res/layout/fragment_crime_list.xml
@@ -1,7 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v7.widget.RecyclerView
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/recycleViewCrimes"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    tools:listitem="@layout/list_item_crime"
     />
