+++ app/src/main/res/layout/list_item_crime_police.xml
@@ -11,7 +11,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/policeRequired"
+        android:text="@string/police_is_required"
         android:textColor="@android:color/holo_red_dark"
         />
