+++ app/src/main/res/values/strings.xml
@@ -4,5 +4,5 @@
     <string name="title">Title</string>
     <string name="detail">Details</string>
     <string name="solved">Solved</string>
-    <string name="policeRequired">POLICE IS REQUIRED!!!!!</string>
+    <string name="police_is_required">POLICE IS REQUIRED!!!!!</string>
 </resources>
